refactor: config モジュールを責務分離する
config.rs を mod/cors/environment に分割し、既存の公開 API を維持したまま routes 側の呼び出しを更新。\n\nCo-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
